+++ .gitignore
@@ -1,5 +1,7 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
+README.md
+
 # dependencies
 /node_modules
 /.pnp
@@ -43,3 +45,5 @@ yarn-error.log*
 
 # IDEA
 /.idea
+
+
