+++ messages/common/en.json
@@ -153,11 +153,12 @@
   "LoginRegister": {
     "login": "ورود",
     "loginToTiara": "ورود به تیــارا",
-    "loginDesc": "برای ورود و استفاده از خدمات تیارا شماره و رمز خود را وارد کنید",
+    "loginDesc": "برای ورود و استفاده از خدمات تیارا نام کاربری و رمز خود را وارد کنید",
     "logout": "Logout",
     "register": "Register",
     "submit": "Submit",
-    "username": "Username",
+    "username": "نام کاربری",
+    "username_placeholder": "نام کاربری خود را وارد کنید",
     "name": "Name",
     "shipping_address": "Shipping Address",
     "phone": "Phone",
